+++ memory/bitmanipulation_amd64.go
@@ -3,7 +3,7 @@ package memory
 // NextPowerOfTwo takes a positive integer and finds the next greater number
 // that's a power of two.  Code directly copied from
 // http://graphics.stanford.edu/~seander/bithacks.html#RoundUpPowerOf2
-func NextPowerOfTwo(size uint) uint {
+func NextPowerOfTwo(size int) int {
 	size--
 	size |= size >> 1
 	size |= size >> 2
@@ -19,10 +19,10 @@ func NextPowerOfTwo(size uint) uint {
 // 8 is passed in, 3 it returned, because 2^3 is 8.  Behavior is undefined if
 // value is not a positive power-of-two.  Code largely inspired by
 // http://graphics.stanford.edu/~seander/bithacks.html#CountBitsSetKernighan
-func PowerOf(value uint) uint {
+func PowerOf(value int) uint8 {
 	v := value - 1
 
-	c := uint(0) // c accumulates the total bits set in v
+	c := uint8(0) // c accumulates the total bits set in v
 	for ; v != 0; c++ {
 		v &= v - 1 // clear the least significant bit set
 	}
